Rename pki selector files to end with -selectors
