+++ ui/tests/acceptance/pki/pki-action-forms-test.js
@@ -14,8 +14,8 @@ import logout from 'vault/tests/pages/logout';
 import enablePage from 'vault/tests/pages/settings/mount-secret-backend';
 import { runCmd } from 'vault/tests/helpers/commands';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
-import { PKI_CONFIGURE_CREATE } from 'vault/tests/helpers/components/pki/pki-configure-create';
-import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root';
+import { PKI_CONFIGURE_CREATE } from 'vault/tests/helpers/components/pki/pki-configure-create-selectors';
+import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root-selectors';
 import { CERTIFICATES } from 'vault/tests/helpers/pki/pki-workflow-helpers';
 
 const { issuerPemBundle } = CERTIFICATES;
