+++ ui/tests/acceptance/pki/pki-configuration-test.js
@@ -16,10 +16,10 @@ import enablePage from 'vault/tests/pages/settings/mount-secret-backend';
 import { runCmd } from 'vault/tests/helpers/commands';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
 import { CERTIFICATES } from 'vault/tests/helpers/pki/pki-workflow-helpers';
-import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root';
-import { PKI_CONFIGURE_CREATE } from 'vault/tests/helpers/components/pki/pki-configure-create';
-import { PKI_DELETE_ALL_ISSUERS } from 'vault/tests/helpers/components/pki/pki-delete-all-issuers';
-import { PKI_ISSUER_LIST } from 'vault/tests/helpers/components/pki/page/pki-issuer-list';
+import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root-selectors';
+import { PKI_CONFIGURE_CREATE } from 'vault/tests/helpers/components/pki/pki-configure-create-selectors';
+import { PKI_DELETE_ALL_ISSUERS } from 'vault/tests/helpers/components/pki/pki-delete-all-issuers-selectors';
+import { PKI_ISSUER_LIST } from 'vault/tests/helpers/components/pki/page/pki-issuer-list-selectors';
 
 const { issuerPemBundle } = CERTIFICATES;
 module('Acceptance | pki configuration test', function (hooks) {
