+++ ui/tests/acceptance/pki/pki-overview-test.js
@@ -11,7 +11,7 @@ import authPage from 'vault/tests/pages/auth';
 import logout from 'vault/tests/pages/logout';
 import enablePage from 'vault/tests/pages/settings/mount-secret-backend';
 import { click, currentURL, currentRouteName, visit } from '@ember/test-helpers';
-import { PKI_OVERVIEW } from 'vault/tests/helpers/components/pki/overview';
+import { PKI_OVERVIEW } from 'vault/tests/helpers/components/pki/overview-selectors';
 import { clearRecords } from 'vault/tests/helpers/pki/pki-workflow-helpers';
 import { runCmd, tokenWithPolicyCmd } from 'vault/tests/helpers/commands';
