+++ ui/tests/helpers/components/pki/pki-configure-create-selectors.ts
