+++ ui/tests/helpers/components/pki/pki-generate-root-selectors.ts
