+++ ui/tests/integration/components/pki/pki-generate-root-test.js
@@ -11,7 +11,7 @@ import { setupEngine } from 'ember-engines/test-support';
 import { setupMirage } from 'ember-cli-mirage/test-support';
 import Sinon from 'sinon';
 import { allowAllCapabilitiesStub } from 'vault/tests/helpers/stubs';
-import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root';
+import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root-selectors';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
 
 module('Integration | Component | pki-generate-root', function (hooks) {
