+++ ui/tests/integration/components/pki/pki-key-form-test.js
@@ -9,7 +9,7 @@ import { render, click, fillIn } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupEngine } from 'ember-engines/test-support';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
-import { PKI_KEY_FORM } from 'vault/tests/helpers/components/pki/pki-key-form';
+import { PKI_KEY_FORM } from 'vault/tests/helpers/components/pki/pki-key-form-selectors';
 import { setupMirage } from 'ember-cli-mirage/test-support';
 import sinon from 'sinon';
