+++ ui/tests/integration/components/pki/pki-key-usage-test.js
@@ -8,7 +8,7 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render, click, findAll } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupEngine } from 'ember-engines/test-support';
-import { PKI_ROLE_FORM } from 'vault/tests/helpers/components/pki/pki-role-form';
+import { PKI_ROLE_FORM } from 'vault/tests/helpers/components/pki/pki-role-form-selectors';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
 
 module('Integration | Component | pki-key-usage', function (hooks) {
