+++ ui/tests/integration/components/pki/pki-role-form-test.js
@@ -9,7 +9,7 @@ import { render, click, fillIn } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupEngine } from 'ember-engines/test-support';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
-import { PKI_ROLE_FORM } from 'vault/tests/helpers/components/pki/pki-role-form';
+import { PKI_ROLE_FORM } from 'vault/tests/helpers/components/pki/pki-role-form-selectors';
 import { setupMirage } from 'ember-cli-mirage/test-support';
 import sinon from 'sinon';
 import { setRunOptions } from 'ember-a11y-testing/test-support';
